+++ library/core/src/main/java/com/google/android/exoplayer2/text/subrip/SubripDecoder.java
@@ -41,8 +41,8 @@ public final class SubripDecoder extends SimpleSubtitleDecoder {
 
   private static final String TAG = "SubripDecoder";
 
-  // Some SRT files don't include hours in the timecode, so we use an optional group.
-  private static final String SUBRIP_TIMECODE = "(?:(\\d+):)?(\\d+):(\\d+),(\\d+)";
+  // Some SRT files don't include hours or milliseconds in the timecode, so we use optional groups.
+  private static final String SUBRIP_TIMECODE = "(?:(\\d+):)?(\\d+):(\\d+)(?:,(\\d+))?";
   private static final Pattern SUBRIP_TIMING_LINE =
       Pattern.compile("\\s*(" + SUBRIP_TIMECODE + ")\\s*-->\\s*(" + SUBRIP_TIMECODE + ")\\s*");
 
@@ -235,7 +235,10 @@ public final class SubripDecoder extends SimpleSubtitleDecoder {
     long timestampMs = hours != null ? Long.parseLong(hours) * 60 * 60 * 1000 : 0;
     timestampMs += Long.parseLong(matcher.group(groupOffset + 2)) * 60 * 1000;
     timestampMs += Long.parseLong(matcher.group(groupOffset + 3)) * 1000;
-    timestampMs += Long.parseLong(matcher.group(groupOffset + 4));
+    @Nullable String millis = matcher.group(groupOffset + 4);
+    if (millis != null) {
+      timestampMs += Long.parseLong(millis);
+    }
     return timestampMs * 1000;
   }